+++ src/test/java/features/WithdrawMoneyFeature.java
@@ -6,6 +6,8 @@ import cucumber.api.java.en.Given;
 import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
 
+import java.math.BigDecimal;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
 
@@ -15,17 +17,17 @@ public class WithdrawMoneyFeature {
     private Account account = new Account(clock);
 
     @Given("^an existing client with \"([^\"]*)\" EUR in his account$")
-    public void initAccountWith(double amount) throws Throwable {
+    public void initAccountWith(BigDecimal amount) throws Throwable {
         account.deposit(amount);
     }
 
     @When("^he withdraws \"([^\"]*)\" EUR form his account$")
-    public void withdrawMoney(double amount) throws Throwable {
+    public void withdrawMoney(BigDecimal amount) throws Throwable {
         account.withdraw(amount);
     }
 
     @Then("^his new balance after withdraws is \"([^\"]*)\" EUR$")
-    public void verifyRunningBalance(double amount) throws Throwable {
+    public void verifyRunningBalance(BigDecimal amount) throws Throwable {
         assertThat(account.runningBalance()).isEqualTo(amount);
     }
 }
